Update 25.3.md
In previous version, second inequality is wrong. It is reversed but it shouldnt and it does not make sense.
I couldnt write the solution here but solved on paper. You can see it here. 
https://www.dropbox.com/s/7v9ef2ixqkshubq/1%2A8rVjzUGtfyD5e2sx-p03Gw.png?dl=0
hope you can write it in an appropriate way in terms of markdown language :)
I will contribute in different questions as well. I have solved and going to share with you.